BAH-4869 sonnar fixes
